annotations

+++ restaurante/src/main/java/com/restaurante/restaurante/repository/RestaurantRepository.java
@@ -3,5 +3,5 @@ package com.restaurante.restaurante.repository;
 import com.restaurante.restaurante.domain.Restaurant;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RestaurantRepository extends JpaRepository<Restaurant,Integer> {
+public interface RestaurantRepository extends JpaRepository<Restaurant,Long> {
 }
